# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CIPHER Intelligence Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: getQuestions
      tags: [gameplay]
      summary: Get questions for current session
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /questions/{questionId}/answer:
    post:
      operationId: submitAnswer
      tags: [gameplay]
      summary: Submit an answer to a question
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
...
  /matches:
    get:
      operationId: getMatches
      tags: [gameplay]
      summary: Get recent matches
      responses:
        "200":
          description: Match list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"

    post:
      operationId: createMatch
      tags: [gameplay]
      summary: Create a new match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{matchId}:
    get:
      operationId: getMatch
      tags: [gameplay]
      summary: Get match details
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{matchId}/complete:
    post:
      operationId: completeMatch
      tags: [gameplay]
      summary: Complete a match and calculate results
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchResultInput"
      responses:
        "200":
          description: Final result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
  /xp/award:
    post:
      operationId: awardXp
      tags: [gameplay]
      summary: Award XP for an action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XpAwardInput"
      responses:
        "200":
          description: XP awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XpAwardResult"
...
  /rankings/global:
    get:
      operationId: getGlobalLeaderboard
      tags: [ranking]
      summary: Get global leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
  /social/friends/request:
    post:
      operationId: sendFriendRequest
      tags: [social]
      summary: Send a friend request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestInput"
      responses:
        "200":
          description: Request sent

  # ─── AI ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Question:
      type: object
      required: [id, type, questionText, difficulty, category, options]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [text, image, video, audio, logic]
        questionText:
          type: string
        difficulty:
          type: integer
          minimum: 1
          maximum: 10
        category:
          type: string
        mediaUrl:
          type: ["string", "null"]
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        timeLimit:
          type: integer
    QuestionOption:
      type: object
      required: [id, text]
      properties:
        id:
          type: integer
        text:
          type: string
    AnswerInput:
      type: object
      required: [optionId, timeSpentMs]
      properties:
        optionId:
          type: integer
        timeSpentMs:
          type: integer
    AnswerResult:
      type: object
      required: [correct, xpGained, correctOptionId, explanation]
      properties:
        correct:
          type: boolean
        xpGained:
          type: integer
        correctOptionId:
          type: integer
        explanation:
          type: string
        streakBonus:
          type: integer
        newLevel:
          type: ["integer", "null"]
...
    XpAwardInput:
      type: object
      required: [action, amount]
      properties:
        action:
          type: string
        amount:
          type: integer
        multiplier:
          type: number
    XpAwardResult:
      type: object
      required: [xpGained, totalXp, newLevel, leveledUp]
      properties:
        xpGained:
          type: integer
        totalXp:
          type: integer
        newLevel:
          type: integer
        leveledUp:
          type: boolean
        unlocks:
          type: array
          items:
            type: string

    # ─── MATCHES ─────────────────────────────────
    Match:
      type: object
      required: [id, type, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [pvp, story, coop, ranked, survival]
        status:
          type: string
          enum: [waiting, active, finished]
        players:
          type: array
          items:
            $ref: "#/components/schemas/MatchPlayer"
        createdAt:
          type: string
          format: date-time
    MatchPlayer:
      type: object
      required: [userId, username, score, isWinner]
      properties:
        userId:
          type: integer
        username:
          type: string
        score:
          type: integer
        rankChange:
          type: integer
        isWinner:
          type: boolean
    MatchInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [pvp, story, coop, ranked, survival]
        opponentId:
          type: ["integer", "null"]
    MatchResultInput:
      type: object
      required: [score, correctAnswers, totalQuestions, timeMs]
      properties:
        score:
          type: integer
        correctAnswers:
          type: integer
        totalQuestions:
          type: integer
        timeMs:
          type: integer
    MatchResult:
      type: object
      required: [matchId, isWinner, xpGained, rankChange, newRankTier]
      properties:
        matchId:
          type: integer
        isWinner:
          type: boolean
        xpGained:
          type: integer
        rankChange:
          type: integer
        newRankTier:
          type: string

    # ─── ACHIEVEMENTS ─────────────────────────────
...
    LeaderboardEntry:
      type: object
      required: [rank, userId, username, avatarUrl, xp, level, rankTier, score]
      properties:
        rank:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        xp:
          type: integer
        level:
          type: integer
        rankTier:
          type: string
        score:
          type: integer
...
    FriendRequestInput:
      type: object
      required: [targetUsername]
      properties:
        targetUsername:
          type: string

    # ─── AI ──────────────────────────────────────
...