# lib/api-zod/src/generated/types/matchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CIPHER Intelligence Game API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchInputType } from './matchInputType';

export interface MatchInput {
  type: MatchInputType;
  /** @nullable */
  opponentId?: number | null;
}